fixed calculator and started working on comparator

+++ cuota/importers/import_tax_data.py
@@ -25,7 +25,7 @@ def get_social_security_bands(path: Path = data_path.joinpath(fn2), annualized: 
     data = pd.read_csv(path, header=0, sep=",")
     for row in data.itertuples(index=False):
         a, b, c = list(row)
-    bands = [Band(floor=floor * x, ceiling=ceiling * x, flat_charge=flat_charge * x)
+    bands = [Band(floor=int(floor * x), ceiling=int(ceiling * x), flat_charge=int(flat_charge * x))
              for floor, ceiling, flat_charge in data.itertuples(index=False)]
     return BandsGroup(bands=bands, year=2025, name="Social Security")
